--- Archives/src/camera_live_view.py
@@ -1,24 +1,0 @@
-import cv2
-from picamera import PiCamera
-from picamera.array import PiRGBArray
-
-camera = PiCamera()
-camera.resolution = (1280,720)
-camera.framerate = 30
-rawCapture = PiRGBArray(camera, size=(640,480))
-
-
-
-
-for frame in camera.capture_continuous(rawCapture, format="bgr", use_video_port=True):
-	# grab the raw NumPy array representing the image, then initialize the timestamp
-	# and occupied/unoccupied text
-	image = frame.array
-	# show the frame
-	cv2.imshow("Frame", image)
-	key = cv2.waitKey(1) & 0xFF
-	# clear the stream in preparation for the next frame
-	rawCapture.truncate(0)
-	# if the `q` key was pressed, break from the loop
-	if key == ord("q"):
-		break
