--- Archives/src/globalvariable.py
@@ -1,3 +1,0 @@
-x_obj = 640
-y_obj = 480
-baseline = 978.744780226674
